+++ packages/client/src/vscode/services.ts
@@ -3,9 +3,9 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import * as monaco from 'monaco-editor';
+import * as monaco from '@codingame/monaco-vscode-editor-api';
 import 'vscode/localExtensionHost';
-import { initialize, IWorkbenchConstructionOptions } from 'vscode/services';
+import { initialize, IWorkbenchConstructionOptions } from '@codingame/monaco-vscode-api';
 import type { OpenEditor } from '@codingame/monaco-vscode-editor-service-override';
 import type { WorkerConfig } from '@codingame/monaco-vscode-extensions-service-override';
 import getExtensionServiceOverride from '@codingame/monaco-vscode-extensions-service-override';
@@ -13,10 +13,10 @@ import getLanguagesServiceOverride from '@codingame/monaco-vscode-languages-serv
 import getModelServiceOverride from '@codingame/monaco-vscode-model-service-override';
 import getLogServiceOverride from '@codingame/monaco-vscode-log-service-override';
 import type { LocalizationOptions } from '@codingame/monaco-vscode-localization-service-override';
-import { EnvironmentOverride } from 'vscode/workbench';
+import { EnvironmentOverride } from '@codingame/monaco-vscode-api/workbench';
 import { Logger } from 'monaco-languageclient/tools';
 import { FakeWorker as Worker } from './fakeWorker.js';
-import { setUnexpectedErrorHandler } from 'vscode/monaco';
+import { setUnexpectedErrorHandler } from '@codingame/monaco-vscode-api/monaco';
 import { initUserConfiguration } from '@codingame/monaco-vscode-configuration-service-override';
 
 export interface MonacoEnvironmentEnhanced extends monaco.Environment {
@@ -132,6 +132,7 @@ export const initServices = async (vscodeApiConfig: VscodeApiConfig, instruction
  *  - quickAccess
  * monaco-languageclient always adds the following services:
  *   - languages
+ *   - log
  *   - model
  */
 export const importAllServices = async (vscodeApiConfig: VscodeApiConfig, instructions?: InitServicesInstructions) => {
@@ -161,7 +162,7 @@ export const importAllServices = async (vscodeApiConfig: VscodeApiConfig, instru
  */
 export const configureExtHostWorker = async (enableExtHostWorker: boolean, userServices: monaco.editor.IEditorOverrideServices) => {
     if (enableExtHostWorker) {
-        const fakeWorker = new Worker(new URL('vscode/workers/extensionHost.worker', import.meta.url), { type: 'module' });
+        const fakeWorker = new Worker(new URL('@codingame/monaco-vscode-api/workers/extensionHost.worker', import.meta.url), { type: 'module' });
         const workerConfig: WorkerConfig = {
             url: fakeWorker.url.toString(),
             options: fakeWorker.options