+++ packages/wrapper/src/vscode/services.ts
@@ -4,9 +4,9 @@
  * ------------------------------------------------------------------------------------------ */
 
 import * as vscode from 'vscode';
-import * as monaco from 'monaco-editor';
+import * as monaco from '@codingame/monaco-vscode-editor-api';
 import { OpenEditor } from '@codingame/monaco-vscode-editor-service-override';
-import { LogLevel } from 'vscode/services';
+import { LogLevel } from '@codingame/monaco-vscode-api';
 import { mergeServices, ViewsConfig, VscodeApiConfig } from 'monaco-languageclient/vscode/services';
 
 export type OverallConfigType = 'extended' | 'classic';