+++ packages/wrapper/src/wrapper.ts
@@ -3,10 +3,10 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import * as monaco from 'monaco-editor';
-import { DisposableStore } from 'vscode/monaco';
-import { LogLevel } from 'vscode/services';
-import { registerExtension, IExtensionManifest, ExtensionHostKind, RegisterExtensionResult } from 'vscode/extensions';
+import * as monaco from '@codingame/monaco-vscode-editor-api';
+import { DisposableStore } from '@codingame/monaco-vscode-api/monaco';
+import { LogLevel } from '@codingame/monaco-vscode-api';
+import { registerExtension, IExtensionManifest, ExtensionHostKind, RegisterExtensionResult } from '@codingame/monaco-vscode-api/extensions';
 import { MonacoLanguageClient } from 'monaco-languageclient';
 import { initServices, InitServicesInstructions, VscodeApiConfig } from 'monaco-languageclient/vscode/services';
 import { Logger, ConsoleLogger } from 'monaco-languageclient/tools';