+++ vite.config.ts
@@ -62,6 +62,8 @@ export const definedViteConfig = defineConfig({
             ]
         },
         include: [
+            '@testing-library/react',
+            'vscode/localExtensionHost',
             'vscode-textmate',
             'vscode-oniguruma'
         ]