+++ packages/examples/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this npm module are documented in this file.
 
+## [2025.1.8] - 2025-01-28
+
+- Updated to `monaco-languageclient@9.2.0-next.5`, `monaco-editor-wrapper@6.2.0-next.5` and `@typefox/monaco-editor-react@6.2.0-next.5`.
+
+## [2025.1.7] - 2025-01-28
+
+- Updated to `monaco-languageclient@9.2.0-next.4`, `monaco-editor-wrapper@6.2.0-next.4` and `@typefox/monaco-editor-react@6.2.0-next.4`.
+
+## [2025.1.6] - 2025-01-27
+
+- Updated to `monaco-languageclient@9.2.0-next.3`, `monaco-editor-wrapper@6.2.0-next.3` and `@typefox/monaco-editor-react@6.2.0-next.3`.
+
+## [2025.1.5] - 2025-01-27
+
+- Updated to `monaco-languageclient@9.2.0-next.2`, `monaco-editor-wrapper@6.2.0-next.2` and `@typefox/monaco-editor-react@6.2.0-next.2`.
+
+## [2025.1.4] - 2025-01-27
+
+- Updated to `monaco-languageclient@9.2.0-next.1`, `monaco-editor-wrapper@6.2.0-next.1` and `@typefox/monaco-editor-react@6.2.0-next.1`.
+
+## [2025.1.3] - 2025-01-24
+
+- Updated to `monaco-languageclient@9.2.0-next.0`, `monaco-editor-wrapper@6.2.0-next.0` and `@typefox/monaco-editor-react@6.2.0-next.0`.
+
 ## [2025.1.2] - 2025-01-20
 
 - Updated to `monaco-languageclient@9.1.1`, `monaco-editor-wrapper@6.1.1` and `@typefox/monaco-editor-react@6.1.1`.