+++ packages/wrapper/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/package/monaco-editor-wrapper) are documented in this file.
 
+## [6.2.0-next.5] - 2025-01-28
+
+- Updated to `monaco-languageclient@9.2.0-next.5`. Updated all `@codingame/monaco-vscode` packages to `13.1.1`.
+
 ## [6.1.1] - 2025-01-20
 
 - View service related imports are made dynamically [#829](https://github.com/TypeFox/monaco-languageclient/pull/829)